add groups table for linux
