+++ siquery_tables/build.rs
@@ -48,6 +48,7 @@ const LINUX_TABLES: &'static [(&'static str)] = &[
     "process_memory_map",
     "process_envs",
     "mounts",
+    "groups"
 ];
 
 const MACOS_TABLES: &'static [(&'static str)] = &[
@@ -60,6 +61,7 @@ const MACOS_TABLES: &'static [(&'static str)] = &[
     "system_info",
     "uptime",
     "mounts",
+    "groups"
 ];
 
 fn emit_features(features: &'static [(&'static str)]) {
