+++ siquery_tables/src/linux/mod.rs
@@ -15,6 +15,7 @@ use serde_json;
 #[cfg(feature = "process_open_sockets")] mod process_open_sockets;
 #[cfg(feature = "processes")] mod processes;
 #[cfg(feature = "mounts")] mod mounts;
+#[cfg(feature = "groups")]          mod groups;
 
 pub struct EtcHostsReader {}
 #[cfg(feature = "etc_hosts")]
