+++ siquery_tables/src/macos/processes.rs
@@ -4,15 +4,39 @@
 
 use libc::*;
 use libc::timeval;
-use std::collections::HashMap;
-use std::mem;
-use std::ptr;
-use tables::ProcessesRow;
 use byteorder::{LittleEndian, ReadBytesExt};
-use std::path::Path;
-use std::ffi::CStr;
-use std::str;
-use std::ffi::OsStr;
+use widestring::WideString;
+use std::{
+    mem,
+    ptr,
+    path::Path,
+    ffi::{
+        CStr,
+        OsStr
+    },
+    str,
+    collections::HashMap
+};
+use winapi::{
+    um::{
+        winnt::PSID,
+        errhandlingapi::GetLastError,
+        winbase::LocalFree
+    },
+    shared::{
+        minwindef::{
+            DWORD,
+            HLOCAL
+        },
+        ntdef::{
+            LPWSTR,
+            NULL
+        },
+        sddl::ConvertSidToStringSidW
+    }
+};
+
+use tables::ProcessesRow;
 
 pub struct ProcCred {
     parent: uint32_t,
@@ -194,6 +218,22 @@ struct rusage_info_v2 {
 }
 
 type mach_timebase_info_data_t = mach_timebase_info;
+
+/// Converts a raw SID into a SID string representation.
+pub fn sid_to_string(sid: PSID) -> Result<String, DWORD> {
+    let mut buf: LPWSTR = NULL as LPWSTR;
+    if unsafe { ConvertSidToStringSidW(sid, &mut buf) } == 0 ||
+        buf == (NULL as LPWSTR) {
+        return Err(unsafe { GetLastError() });
+    }
+
+    let buf_size = unsafe { libc::wcslen(buf) };
+    let sid_string = unsafe { WideString::from_ptr(buf, buf_size) };
+
+    unsafe { LocalFree(buf as HLOCAL) };
+
+    Ok(sid_string.to_string_lossy())
+}
 
 impl ProcessesRow {
     fn new() -> ProcessesRow {
